\name{choosePoints}
\alias{choosePoints}
\title{Choose points from a trellis plot.}
\description{
A interactive tool for selecting a set of points inside a user-defined border.}
\usage{
choosePoints(...)
}
\arguments{
  \item{\dots}{Not used.}
}
\details{A trellis windows must be active.}
\value{A logical vector where the selected points are labeled with the
  TRUE value.}

\author{Oscar Perpiñán Lamigueiro}

\seealso{\code{trellis.focus}, \code{grid.locator}, \code{unit}, \code{in.out}}
