\name{hankel}
\alias{hankel}

\title{
Hankel matrix
}
\description{Creates a hankel matrix from a vector.
}
\usage{
hankel(x, ncol = length(x)\%/\%2)
}

\arguments{
  \item{x}{
A numeric vector.
}
  \item{ncol}{numeric, number of columns of the matrix.
}
}

\value{A matrix.}

\references{
%% ~put references to the literature/web site here ~
}
\author{Oscar Perpiñán Lamigueiro}

\examples{

}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
