\name{isFiltered-methods}
\docType{methods}
\alias{isFiltered-methods}
\alias{isFiltered,PD-method}
\title{Methods for Function \code{isFiltered} }
\description{
Methods for function \code{isFiltered} 
}
\section{Methods}{
\describe{

\item{\code{signature(object = "PD")}}{TRUE if the object has been
  filtered with \code{\link{filterPD}}
}
}}
\keyword{methods}
\author{Oscar Perpiñán Lamigueiro}
