\name{isTransformed-methods}
\docType{methods}
\alias{isTransformed}
\alias{isTransformed-methods}
\alias{isTransformed,PD-method}
\title{Methods for Function \code{isTransformed} }
\description{
Methods for function \code{isTransformed} 
}
\section{Methods}{
\describe{

\item{\code{signature(object = "PD")}}{TRUE if the object has been
  previously tranformed with \code{\link{transformPD-methods}}

}
}}
\keyword{methods}
\author{Oscar Perpiñán Lamigueiro}
