\name{label.densityplot}
\alias{label.densityplot}
\title{Labels for a densityplot
}
\description{Labels for a densityplot
}
\usage{
label.densityplot(x, group.number, col.line)
}
\arguments{
  \item{x}{numeric}
  \item{group.number}{numeric, for \code{\link{glayer}}.}
  \item{col.line}{numeric, for \code{\link{glayer}}.}}

\value{A trellis layer for use with \code{\link{glayer}}}
\author{Oscar Perpiñán Lamigueiro}
\seealso{\code{\link{layer}}, \code{\link{densityplot}}}

