\name{nZeroCross}
\alias{nZeroCross}

\title{Zero-cross rate}
\description{Zero-cross rate}
\usage{nZeroCross(x, ...)}

\arguments{
  \item{x}{numeric}
  \item{\dots}{additional arguments for \code{\link{no0}}
}
}
\value{A numeric.}

\author{Oscar Perpiñán Lamigueiro}
\note{The function uses \code{\link{no0}}}

\seealso{
  \code{\link{no0}}
  }

