\name{PD2Long-methods}
\docType{methods}
\alias{PD2Long-methods}
\alias{PD2Long,PD-method}
\alias{PD2Long,PDCluster-method}
\title{Methods for Function \code{PD2Long} }

\description{
Methods for function \code{PD2Long} 
}
\section{Methods}{
\describe{

\item{\code{signature(object = "PD")}}{
A data.frame with a long version of the \code{data} slot of the object.
}

\item{\code{signature(object = "PDCluster")}}{
A data.frame with a long version of the \code{data} slot of the
object. It includes a column named \code{cluster}.
}
}}
\keyword{methods}
\author{Oscar Perpiñán Lamigueiro}
