\name{plot.prony}
\alias{plot.prony}

\title{
Plot of a Prony decomposition.
}
\description{
%% A concise (1-5 lines) description of what the function does.
}
\usage{
plot.prony(x, all = TRUE, ...)
}

\arguments{
  \item{x}{A \code{\link{prony}} object
}
  \item{all}{logical, should all the components be drawn?}
  \item{\dots}{Additional arguments for \code{\link{prony}}}
}
\details{
%% If necessary, more details than the description above
}
\value{A trellis graphic.}

\author{Oscar Perpiñán Lamigueiro}


\seealso{
\code{\link{xyplot}}, \code{\link{prony}}
}
\examples{
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
