\name{show-methods}
\docType{methods}
\alias{show-methods}
\alias{show,PD-method}
\alias{show,PDCluster-method}
\title{Methods for Function \code{show} in Package \pkg{methods}}
\description{
Methods for function \code{show} in package \pkg{methods}
}
\section{Methods}{
\describe{


\item{\code{signature(object = "PD")}}{
%% describe this method here~~
}

\item{\code{signature(object = "PDCluster")}}{
%% describe this method here~~
}

}}
\keyword{methods}
\author{Oscar Perpiñán Lamigueiro}
