\name{wavVarPD}
\alias{wavVarPD}

\title{Wavelet variance
}
\description{Wavelet variance of a partial discharge signal.}
\usage{
wavVarPD(x, ...)
}
\arguments{
  \item{x}{numeric, the partial discharge signal}
  \item{\dots}{additional arguments for \code{\link{wavVar}}}
}

\value{A numeric vector with the relative wavelet variance for several levels.}

\author{Oscar Perpiñán Lamigueiro}

\seealso{\code{\link{wavVar}}
}
