\name{hankel}
\alias{hankel}

\title{
Hankel matrix
}
\description{Creates a hankel matrix from a vector.
}
\usage{
hankel(x, ncol = length(x)\%/\%2)
}

\arguments{
  \item{x}{
A numeric vector.
}
  \item{ncol}{numeric, number of columns of the matrix.
}
}

\value{A matrix.}


\author{Oscar Perpiñán Lamigueiro}

