\name{nZeroCross}
\alias{nZeroCross}

\title{Zero-cross rate}
\usage{nZeroCross(x, ...)}

\arguments{
  \item{x}{numeric}
  \item{\dots}{additional arguments for \code{\link{no0}}
}
}
\value{A numeric.}
\references{
%% ~put references to the literature/web site here ~
}
\author{Oscar Perpiñán Lamigueiro}
\note{The function uses \code{\link{no0}}}

\seealso{
  \code{\link{no0}}
  }
\examples{
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
