\name{xyplot.prony}
\alias{xyplot.prony}

\title{
Plot of a Prony decomposition.
}
\description{
Plot of a Prony decomposition.
}
\usage{
xyplot.prony(x, data, all = FALSE, ...)
}

\arguments{
  \item{x}{A \code{\link{prony}} object
  }
  \item{data}{Not used.}
  \item{all}{logical, should all the components be drawn?}
  \item{\dots}{Additional arguments for \code{\link{prony}}}
}

\value{A trellis graphic.}

\author{Oscar Perpiñán Lamigueiro}


\seealso{
\code{\link{xyplot}}, \code{\link{prony}}
}

