\name{transformPD-methods}
\docType{methods}
\alias{transformPD-methods}
\alias{transformPD,PD-method}
\title{Methods for Function \code{transformPD} }
\description{
Box-Cox transformations.
}


\section{Methods}{
\describe{

\item{\code{signature(object = "PD")}}{ See \code{\link{powerTransform}}
}
}}
\keyword{methods}
\author{Oscar Perpiñán Lamigueiro}
