\name{wavVarPD}
\alias{wavVarPD}

\title{Wavelet variance
}
\description{Wavelet variance of a partial discharge signal.}
\usage{
wavVarPD(x, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{numeric, the partial discharge signal}
  \item{\dots}{additional arguments for \code{\link{wavVar}}}
}
\details{
%% If necessary, more details than the description above
}
\value{A numeric vector with the relative wavelet variance for several levels.%% ...
}
\references{
%% ~put references to the literature/web site here ~
}
\author{Oscar Perpiñán Lamigueiro}
\note{
%% further notes~~
}

\seealso{\code{\link{wavVar}}
}
\examples{
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
