\name{xyplot-methods}
\docType{methods}
\alias{xyplot-methods}
\alias{xyplot,PDCluster,missing-method}
\alias{xyplot,PD,missing-method}
\title{Methods for Function \code{xyplot} in Package \pkg{pdR}}
\description{
Methods for function \code{xyplot} in package \pkg{pdR}
}
\usage{
\S4method{xyplot}{PD}(x, par.settings, plot.refl, alpha, ...)
\S4method{xyplot}{PDCluster}(x, par.settings, plot.refl, alpha, ...)
}
\arguments{
  \item{plot.refl}{logical, should be the reflections be drawn in
    separate panels?}
  \item{par.settings, alpha, ...}{See \code{\link{xyplot}}}
}
% \section{Methods}{
% \describe{


% \item{\code{signature(x = "PDCluster", data = "missing")}}{
% %% describe this method here~~
% }

% \item{\code{signature(x = "PD", data = "missing")}}{
% %% describe this method here~~
% }

% }}
\keyword{methods}
\author{Oscar Perpiñán Lamigueiro}
